requirements_py36.txt
appdirs==1.4.3
Django==1.9.8
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
PyJWT==1.4.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
pytz==2016.10
six==1.10.0
